# backend/docs/API_DOCUMENTATION.yaml
openapi: 3.0.3
info:
  title: CozyVTT Backend Route Reference
  description: |
    The HTTP routes CozyVTT's own web client calls, described for people working
    on CozyVTT itself.

    ## No stability guarantee
    This is an internal surface, not a public API. It is **not versioned**, it
    carries **no compatibility promise**, and any route here may change shape or
    disappear in a point release without notice or a deprecation period.

    There are no API keys, no bearer tokens and no service accounts. The only
    credential is the session cookie described below — but a program can obtain
    one exactly as the web client does, by posting a user's own email and
    password to `/api/auth/login` and keeping the cookie it returns. That cookie
    also authenticates the Socket.io connection, which shares the same session.

    So scripting CozyVTT is possible, and people have done it. What it does not
    come with is a promise: a script authenticates **as a specific user** and can
    do only what that user could do in the browser, the session expires after an
    hour of inactivity (refreshed by activity), an account with MFA enabled needs
    the second step, and any route here may change shape in a point release
    without notice. Treat this document as a map of the current code rather than
    a contract.

    ## Authentication
    All endpoints (except `/health`, `/api/setup/*`, and public auth routes) require authentication via session cookies.
...

...

...

...

...

...

...

...
servers:
...
security:
  - cookieAuth: []
...
paths:
...
  /api/users:
    get:
      tags:
        - Users
      summary: List all users
      description: Get list of all users (Admin only)
      operationId: listUsers
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Get specific user (own profile or admin)
      operationId: getUserById
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Users
      summary: Update user
      description: Update user profile (own profile or admin). Admin can change platformRole, which ends every login session the target has, because the role is carried in the session. The globalAssetManager and templateEditor flags are read from the database at the point of use, so changing those takes effect on the next request and does not sign anyone out.
      operationId: updateUser
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                email:
                  type: string
                  format: email
                avatarUrl:
                  type: string
                platformRole:
                  type: string
                  enum: [ADMIN, USER]
                  description: Admin only
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Users
      summary: Delete user
      description: Delete user account (Admin only, cannot delete self). Ends every login session the deleted user had.
      operationId: deleteUser
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/users/{id}/reset-password:
    post:
      tags:
        - Users
      summary: Admin reset user password
      description: Generate temporary password for user (Admin only). Sets mustChangePassword flag so user is forced to change on next login.
      operationId: adminResetUserPassword
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Password reset with temporary password
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  temporaryPassword:
                    type: string
                  mustChangePassword:
                    type: boolean
                    example: true
                  notice:
                    type: string
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/users/{id}/preferences:
    get:
      tags:
        - Users
      summary: Get user preferences
      description: |
        Returns the user's per-user preferences blob (theme, font, dice color, sound effects, notifications).
        Returns an empty object if the user has not customized any preferences yet — the frontend layers
        these on top of the system defaults from `GET /api/auth/appearance`.
        Users can fetch their own preferences; admins can fetch any user's.
      operationId: getUserPreferences
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User preferences
          content:
            application/json:
              schema:
                type: object
                properties:
                  preferences:
                    $ref: '#/components/schemas/UserPreferences'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Users
      summary: Update user preferences (partial merge)
      description: |
        Partial-merge update of the user's preferences. Send any subset of fields and the server
        merges them into the stored JSON blob.
        Users can update their own preferences; admins can update any user's.
        Unknown fields are stripped silently. Empty payloads are rejected with 400.
      operationId: updateUserPreferences
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPreferences'
      responses:
        '200':
          description: Preferences updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  preferences:
                    $ref: '#/components/schemas/UserPreferences'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/users/{id}/send-reset-link:
    post:
      tags:
        - Users
      summary: Admin send password reset link
      description: |
        Send a password reset email to the user (Admin only).
        Invalidates any existing unused reset tokens and creates a new one (expires in 1 hour).
        Requires SMTP to be configured — returns 503 if SMTP is not set up.
      operationId: adminSendUserResetLink
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Reset email sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password reset email sent successfully
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          description: SMTP not configured
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Service Unavailable
                  message:
                    type: string
                    example: SMTP is not configured on this instance.

  # ============================================
  # CAMPAIGNS
  # ============================================
...
  /api/admin/users:
    post:
      tags: [Admin]
      summary: Create a new user
      description: |
        Creates a user with a generated temporary password. The account is flagged
        `mustChangePassword`, so that password can only be used to set a real one — every
        other endpoint returns 403 `PASSWORD_CHANGE_REQUIRED` until it is replaced.

        When SMTP is configured the credentials are emailed and `temporaryPassword` is
        **withheld** from the response; it is returned only when the email could not be
        sent, so the admin never holds a working credential unnecessarily. To create an
        account without any password at all, use `/api/admin/users/invite`.
      operationId: adminCreateUser
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                displayName:
                  type: string
                  maxLength: 50
                platformRole:
                  type: string
                  enum: [USER, ADMIN]
                  default: USER
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
                  emailSent:
                    type: boolean
                    description: Whether the welcome email with the credentials was delivered
                  temporaryPassword:
                    type: string
                    description: >-
                      One-time temporary password. Present ONLY when emailSent is false —
                      share it securely; it is not retrievable afterwards.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Email already in use
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cozyvtt.sid
      description: Session cookie obtained via /api/auth/login
...
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        platformRole:
          type: string
          enum: [ADMIN, USER]
        mfaEnabled:
          type: boolean
        avatarUrl:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
    UserPreferences:
      type: object
      description: Per-user appearance preferences (theme + font). All fields optional; partial updates are merged into existing preferences.
      properties:
        themeId:
          type: string
          enum:
            - cozy-default
            - autumn-hearth
            - desert-sand
            - rose-garden
            - ocean-depths
            - northern-frost
            - enchanted-forest
            - twilight-vale
            - obsidian-night
            - shadow-realm
            - deep-dungeon
            - midnight-court
            - parchment-classic
            - slate-modern
            - dragonfire
            - arcane-storm
            - custom
          description: Preset theme identifier, or "custom" to use customThemeColors.
        customThemeColors:
          type: object
          nullable: true
          description: Required when themeId is "custom". Set to null to clear.
          properties:
            primary:
              type: string
              pattern: '^#[0-9a-fA-F]{6}$'
            accent:
              type: string
              pattern: '^#[0-9a-fA-F]{6}$'
            background:
              type: string
              pattern: '^#[0-9a-fA-F]{6}$'
            text:
              type: string
              pattern: '^#[0-9a-fA-F]{6}$'
          required: [primary, accent, background, text]
        fontId:
          type: string
          enum:
            - default
            - medieval
            - elegant
            - modern
            - handwritten
            - clean
            - scholarly
            - gothic
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: BadRequest
          description: Short machine-readable error identifier
        message:
          type: string
          example: One or more fields failed validation.
          description: Human-readable error description
        details:
          type: object
          nullable: true
          additionalProperties: true
          description: Optional structured details (e.g. per-field Zod issues)
...
  responses:
    BadRequest:
      description: Bad Request - Invalid input or validation failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: BadRequest
            message: Required field "name" is missing.
    Unauthorized:
      description: Unauthorized - Not authenticated or MFA pending.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Authentication required.
    Forbidden:
      description: Forbidden - Authenticated but insufficient permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Forbidden
            message: You do not have permission to perform this action.
...
    NotFound:
      description: Not Found - Resource does not exist or you cannot see it.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: NotFound
            message: Campaign not found.
...